requirements.txt
PyYAML==6.0.3
requests==2.20.0
tabulate==0.8.2
urllib3==1.24.2
